factor migration and downgrade into another file
